+++ open/start.go
@@ -3,7 +3,7 @@ package open
 func Start(input string) error {
 	cmd, ok := open(input)
 	if !ok {
-		return unsupportedOSError
+		return errUnsupportedOS
 	}
 
 	return cmd.Start()
@@ -12,7 +12,7 @@ func Start(input string) error {
 func StartWith(input, with string) error {
 	cmd, ok := openWith(input, with)
 	if !ok {
-		return unsupportedOSError
+		return errUnsupportedOS
 	}
 
 	return cmd.Start()